fix(monetary): migrar CashRegisterService y CashMovement a integer
CORRECCIONES CRÍTICAS:

1. CashRegisterService:
   - float $openingAmountDefault = 50000.0 → int $openingAmountDefault = 50000
   - float $maxAmount = 500000.0 → int $maxAmount = 500000

2. CashMovement::balanceImpact():
   - Tipo de retorno: float → int
   - Cast: (float) → (int)
   - return (int) $this->amount * $this->type->balanceSign()

JUSTIFICACIÓN:
Ambos son montos monetarios CLP, no porcentajes ni cantidades físicas.
Deben ser integer para consistencia con ADR-018.

NO TOCADOS (porcentajes):
- TipPolicy.waiter_percentage (decimal:2 correcto)
- TipPolicy.pool_percentage (decimal:2 correcto)

RESULTADO:
✅ Cashier module 100% integer en montos monetarios
✅ Porcentajes mantienen decimal (correcto)

+++ app/Modules/Cashier/Domain/Entities/CashMovement.php
@@ -94,7 +94,7 @@ class CashMovement extends Model
      */
     public function balanceImpact(): int  // ADR-018: CLP entero
     {
-        return (float) $this->amount * $this->type->balanceSign();
+        return (int) $this->amount * $this->type->balanceSign();
     }
 
     public function scopeOfType($query, MovementType $type)
